✨refactor: getNews 메서드 분리
