+++ src/main/java/com/real/backend/domain/news/service/NewsCommentService.java
@@ -9,6 +9,7 @@ import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.real.backend.domain.news.component.NewsFinder;
 import com.real.backend.domain.news.domain.News;
 import com.real.backend.domain.news.dto.NewsCommentRequestDTO;
 import com.real.backend.domain.news.repository.NewsRepository;
@@ -21,7 +22,6 @@ import com.real.backend.domain.news.domain.NewsComment;
 import com.real.backend.domain.news.dto.NewsCommentListResponseDTO;
 import com.real.backend.domain.news.repository.NewsCommentRepository;
 import com.real.backend.domain.user.domain.User;
-import com.real.backend.domain.user.service.UserService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -31,9 +31,8 @@ public class NewsCommentService {
 
     private final NewsCommentRepository newsCommentRepository;
     private final NewsRepository newsRepository;
-    private final UserService userService;
-    private final NewsService newsService;
     private final UserFinder userFinder;
+    private final NewsFinder newsFinder;
 
     public SliceDTO<NewsCommentListResponseDTO> getNewsCommentListByCursor(Long newsId, Long cursorId, String cursorStandard, int limit, Long currentUserId) {
 
@@ -63,8 +62,8 @@ public class NewsCommentService {
             throw new BadRequestException("필수 파라미터인 commentId를 받지 못했습니다.");
         }
 
-        News news = newsService.getNews(newsId);
-        NewsComment newsComment = newsCommentRepository.findById(commentId).orElseThrow(() -> new NotFoundException("해당 id를 가진 뉴스가 존재하지 않습니다."));
+        News news = newsFinder.getNews(newsId);
+        NewsComment newsComment = newsCommentRepository.findById(commentId).orElseThrow(() -> new NotFoundException("해당 id를 가진 댓글이 존재하지 않습니다."));
 
         if (!newsComment.getUser().getId().equals(userId)) {
             throw new ForbiddenException("해당 댓글 작성자가 아닙니다.");
@@ -77,7 +76,7 @@ public class NewsCommentService {
     @Transactional
     public void createNewsComment(Long newsId, Long userId, NewsCommentRequestDTO newsCommentRequestDTO) {
         User user = userFinder.getUser(userId);
-        News news = newsService.getNews(newsId);
+        News news = newsFinder.getNews(newsId);
         news.increaseCommentCount();
 
         newsCommentRepository.save(NewsComment.builder()
